Remove static method from flavor manager

+++ app/services/atmosphere/flavor_manager.rb
@@ -1,53 +1,53 @@
-# This class is used to populate the Atmosphere database with information regarding the available virtual machine flavors for each compute site
+# Populate the Atmosphere database with information regarding the
+# available virtual machine flavors for each compute site.
 module Atmosphere
   class FlavorManager
-
-    # Scan each Tenant and update the Atmosphere database as needed
-    def self.scan_all_tenants
-      Tenant.all.each do |t|
-        Rails.logger.info("Updating virtual machine flavors for tenant #{t.id}.")
-        self.scan_tenant(t)
-      end
+    def initialize(tenant)
+      @tenant = tenant
     end
 
-    def self.scan_tenant(t)
-      begin
-        # Purge flavors which no longer exist in tenant
-        existing_flavors = t.cloud_client.flavors.collect{|f| f.id}
-        t.virtual_machine_flavors.each do |flavor|
-          if !(existing_flavors.include? flavor.id_at_site) and flavor.virtual_machines.count == 0
-            flavor.destroy
+    def execute
+      update_existing_flavors
+      purge_non_existing_flavors
+    rescue Exception => e
+      Rails.logger.error(I18n.t('virtual_machine_flavor.update_flavors_failed',
+                                id: tenant.tenant_id, msg: e.message))
+    end
+
+    private
+
+    attr_reader :tenant
+
+    def purge_non_existing_flavors
+      tenant.virtual_machine_flavors.
+        where.not(id_at_site: cloud_flavors.map(&:id)).each do |flavor|
+          flavor.destroy if flavor.virtual_machines.count == 0
+        end
+    end
+
+    def update_existing_flavors
+      cloud_flavors.each { |flavor| update(flavor) }
+    end
+
+    def cloud_flavors
+      @cloud_flavors ||= tenant.cloud_client.flavors
+    end
+
+    def update(cloud_flavor)
+      tenant.virtual_machine_flavors.
+        find_or_initialize_by(id_at_site: cloud_flavor.id).tap do |flavor|
+          flavor.flavor_name = cloud_flavor.name
+          flavor.cpu = cloud_flavor.vcpus
+          flavor.memory = cloud_flavor.ram
+          flavor.hdd = cloud_flavor.disk
+          flavor.supported_architectures = cloud_flavor.supported_architectures
+
+          unless flavor.save
+            Rails.logger.error(I18n.t('virtual_machine_flavor.update_failed',
+                                      name: cloud_flavor.name,
+                                      error: flavor.errors))
           end
         end
-
-        # Retrieve the current list of flavors from t and update Atmo representation accordingly
-        t.cloud_client.flavors.each do |flavor|
-          self.check_and_update_flavor(t, flavor)
-        end
-      rescue Exception => e
-        Rails.logger.error("Unable to update flavors for tenant #{t.id}: #{e.message}. Skipping.")
-      end
-    end
-
-    # Upserts the selected flavor in the selected tenant. If this flavor is already defined for this tenant, its parameters are updated.
-    # Otherwise a new record is created with hourly cost = 0.
-    def self.check_and_update_flavor(t, flavor)
-      # Check if this flavor is already defined for t. Assuming that the flavor is uniquely identified by its id_at_site
-      vm_flavor = t.virtual_machine_flavors.find_or_initialize_by(id_at_site: flavor.id.to_s)
-      vm_flavor.flavor_name = flavor.name
-      vm_flavor.cpu = flavor.vcpus
-      vm_flavor.memory = flavor.ram
-      vm_flavor.hdd = flavor.disk
-      vm_flavor.supported_architectures = flavor.supported_architectures
-
-      unless vm_flavor.save
-        Rails.logger.error("Unable to save vm flavor with name #{vm_flavor.flavor_name}: nested exception is #{vm_flavor.errors}")
-      end
-    end
-
-    # Checks if the flavor with a given ID is defined in tenant t
-    def self.exists_in_tenant? (t, flavor_id)
-      t.cloud_client.flavors.select{|f| f.id.to_s == flavor_id.to_s}.count > 0
     end
   end
 end
